refactor: dto간 상속 관계 제거
- 각각의 dto간의 의존 관계 생기지 않도록 상속 관계 제거
